Move new helper method to new file and class
