+++ administrator/components/com_installer/views/database/tmpl/default.php
@@ -33,11 +33,11 @@ defined('_JEXEC') or die;
 						<?php endif; ?>
 
 						<?php if ($this->schemaVersion != $this->changeSet->getSchema()) : ?>
-							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_SCHEMA_ERROR', $this->schemaVersionHtml, JHtml::_('inlineBidirectional', $this->changeSet->getSchema(), 'ltr')); ?></li>
+							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_SCHEMA_ERROR', $this->schemaVersionHtml, JHtml::_('language.inlineBidirectional', $this->changeSet->getSchema(), 'ltr')); ?></li>
 						<?php endif; ?>
 
 						<?php if (version_compare($this->updateVersion, JVERSION) != 0) : ?>
-							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_UPDATEVERSION_ERROR', $this->updateVersionHtml, JHtml::_('inlineBidirectional', JVERSION, 'ltr')); ?></li>
+							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_UPDATEVERSION_ERROR', $this->updateVersionHtml, JHtml::_('language.inlineBidirectional', JVERSION, 'ltr')); ?></li>
 						<?php endif; ?>
 
 						<?php foreach ($this->errors as $line => $error) : ?>
@@ -58,8 +58,8 @@ defined('_JEXEC') or die;
 				<div class="control-group" >
 					<fieldset class="panelform">
 						<ul>
-							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_SCHEMA_VERSION_IN_TABLE', JHtml::_('inlineBidirectional', '#__schemas', 'ltr'), $this->schemaVersionHtml); ?></li>
-							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_UPDATE_VERSION_IN_TABLE', JHtml::_('inlineBidirectional', '#__extensions', 'ltr'), $this->updateVersionHtml); ?></li>
+							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_SCHEMA_VERSION_IN_TABLE', JHtml::_('language.inlineBidirectional', '#__schemas', 'ltr'), $this->schemaVersionHtml); ?></li>
+							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_UPDATE_VERSION_IN_TABLE', JHtml::_('language.inlineBidirectional', '#__extensions', 'ltr'), $this->updateVersionHtml); ?></li>
 							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_DRIVER', JFactory::getDbo()->name); ?></li>
 							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_CHECKED_OK', count($this->results['ok'])); ?></li>
 							<li><?php echo JText::sprintf('COM_INSTALLER_MSG_DATABASE_SKIPPED', count($this->results['skipped'])); ?></li>
