+++ administrator/components/com_installer/views/update/tmpl/default.php
@@ -105,10 +105,10 @@ $listDirn  = $this->escape($this->state->get('list.direction'));
 									<?php echo $item->type_translated; ?>
 								</td>
 								<td class="hidden-phone center">
-									<span class="label label-warning"><?php echo JHtml::_('inlineBidirectional', $item->current_version, 'ltr'); ?></span>
+									<span class="label label-warning"><?php echo JHtml::_('language.inlineBidirectional', $item->current_version, 'ltr'); ?></span>
 								</td>
 								<td class="center">
-									<span class="label label-success"><?php echo JHtml::_('inlineBidirectional', $item->version, 'ltr'); ?></span>
+									<span class="label label-success"><?php echo JHtml::_('language.inlineBidirectional', $item->version, 'ltr'); ?></span>
 								</td>
 								<td class="hidden-phone center">
 									<?php echo $item->folder_translated; ?>
